+++ src-app/change_log.md
@@ -1,3 +1,8 @@
+SynDaver Symple Slicer v1.0.19
+==============================
+
+* '''Fimrware:''' Disable wireless "Update" button when wireless printer is unavailable.
+
 SynDaver Symple Slicer v1.0.18
 ==============================